refactor testimonials path by adding controller folder
